Now the clients will get a default primary Savings Account.

+++ HermesBanking.Core.Application/DTOs/User/SaveUserDto.cs
@@ -10,7 +10,7 @@
         public required string Password { get; set; }
         public required string Role { get; set; }
         public required string UserId { get; set; }
-        public double? InitialAmount { get; set; }
+        public decimal? InitialAmount { get; set; }
         public required bool IsActive { get; set; }
     }
 }
