+++ HermesBanking.Core.Application/DTOs/User/UserDto.cs
@@ -10,7 +10,7 @@
         public bool? isVerified { get; set; }
         public required string Role { get; set; }
         public required string UserId { get; set; }
-        public double? InitialAmount { get; set; }
+        public decimal? InitialAmount { get; set; }
         public required bool IsActive { get; set; }
     }
 }
