+++ HermesBanking.Core.Application/ViewModels/User/ToggleUserStateViewModel.cs
@@ -10,7 +10,7 @@
         public string? Password { get; set; }
         public string? Role { get; set; }
         public string? UserId { get; set; }
-        public double? InitialAmount { get; set; }
+        public decimal? InitialAmount { get; set; }
         public required bool IsActive { get; set; }
     }
 }
